Refaktorert manuell jfr komponent.
